[refactor/#13] refactor: 도메인에 따른 패키지 구조 변경
