+++ src/main/java/codesquad/bows/project/dto/ProjectDetailResponse.java
@@ -1,4 +1,4 @@
-package codesquad.bows.dto;
+package codesquad.bows.project.dto;
 
 import java.time.LocalDateTime;
 import java.util.List;
