+++ src/main/java/codesquad/bows/project/dto/ServiceMetadata.java
@@ -1,4 +1,4 @@
-package codesquad.bows.dto;
+package codesquad.bows.project.dto;
 
 import io.kubernetes.client.openapi.models.V1Service;
 import io.kubernetes.client.openapi.models.V1ServicePort;
