+++ src/main/java/codesquad/bows/project/repository/ProjectRepository.java
@@ -1,7 +1,7 @@
-package codesquad.bows.repository;
+package codesquad.bows.project.repository;
 
-import codesquad.bows.dto.ProjectMetadata;
-import codesquad.bows.entity.Project;
+import codesquad.bows.project.dto.ProjectMetadata;
+import codesquad.bows.project.entity.Project;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -11,14 +11,14 @@ import java.util.List;
 public interface ProjectRepository extends JpaRepository<Project, Long> {
 
     @Query(""" 
-           SELECT new codesquad.bows.dto.ProjectMetadata(p.projectName, p.createdAt, p.domain)
+           SELECT new codesquad.bows.project.dto.ProjectMetadata(p.projectName, p.createdAt, p.domain)
            FROM Project p
            WHERE p.id = :id
            """)
     ProjectMetadata getMetadataById(@Param("id") Long id);
 
     @Query(""" 
-           SELECT new codesquad.bows.dto.ProjectMetadata(p.projectName, p.createdAt, p.domain)
+           SELECT new codesquad.bows.project.dto.ProjectMetadata(p.projectName, p.createdAt, p.domain)
            FROM Project p
            """)
     List<ProjectMetadata> findAllProjectMetadata();
