+++ src/main/java/codesquad/bows/project/service/BashExecutor.java
@@ -1,4 +1,4 @@
-package codesquad.bows.service;
+package codesquad.bows.project.service;
 
 import lombok.extern.slf4j.Slf4j;
