+++ src/main/resources/application-prod.yaml
@@ -1,0 +1,36 @@
+spring:
+  config:
+    activate:
+      on-profile: prod
+
+  application:
+    name: bows
+  datasource:
+    url: jdbc:mysql://localhost:3306/bows
+    username: bows_dev
+    password: qwer1234
+    driver-class-name: com.mysql.cj.jdbc.Driver
+  sql:
+    init:
+      mode: always
+
+  jpa:
+    hibernate:
+      naming:
+        physical-strategy: org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
+      ddl-auto: none
+    properties:
+      hibernate:
+        show_sql: true
+        format_sql: true
+    database-platform: org.hibernate.dialect.MySQLDialect
+    open-in-view: false
+
+cors:
+  originUrl: http://bows.co.kr:81
+
+kube:
+  configPath: "/.kube/config"
+  helm:
+    repoName: "simpleApp"
+    chartName: "bows-simpleApp"
