+++ src/main/resources/data.sql
@@ -1,0 +1,4 @@
+INSERT INTO Project
+(PROJECT_NAME, DOMAIN, BACKEND_IMAGE_NAME, FRONTEND_IMAGE_NAME, DB_PASSWORD, DB_ENDPOINT, DB_USER_NAME, DB_USER_PASSWORD)
+VALUES
+    ('example', 'alpha.example.com', 'bogyumkim/backend', 'bogyumkim/frontend', 'password123', 'db.alpha.example.com', 'db_user_alpha', 'user_password_alpha');
